New features in typechecking lang to facilitate working with types
Introduce TypeTermTable as an alternative to TermTable dedicated to
types. A "type term" declaration has a type builder: an expression
which creates the corresponding SNode.
To access term's features on can use FeatureReferenceExpression.

The new _ExpectType_ _RecoverType_ rule tables are automatically added
to the user's 'types' aspect and serve to extract types from types.

AssignType and RecoverType pseudo-constraints can be used to
extract/assign or just extract the type, correspondingly.